requirements-minimal.txt
streamlit==1.32.0
pandas==1.5.3
numpy==1.24.0
plotly==5.18.0
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0 